modularized shiftController, added console.logs for employeeController and shiftController middleware
